+++ src/POST_auth0_change_password.d.ts
@@ -1,0 +1,6 @@
+/// <reference lib="dom" />
+import type { Ctx } from '@ctx-core/object'
+export declare function POST_auth0_change_password(ctx:Ctx, req:Request):Promise<Response>
+export interface POST_auth0_change_password_body_T {
+	password:string
+}