+++ src/jwt_token_decoded_.d.ts
@@ -1,3 +1,0 @@
-import type { jwt_token_decoded_I } from '@ctx-core/auth0'
-import type { Ctx } from '@ctx-core/object'
-export declare function jwt_token_decoded_(ctx:Ctx, jwt_token:string):Promise<jwt_token_decoded_I>