+++ src/verify_jwt_user_id.js
@@ -1,5 +1,5 @@
 import { user_id_ } from '@ctx-core/auth0'
-import { koa_jwt_token_decoded_ } from './koa_jwt_token_decoded_.js'
+import { strict_verify_jwt_token } from './strict_verify_jwt_token.js'
 /**
  * @param {import('@ctx-core/object').Ctx}ctx
  * @param {string}authorization
@@ -7,6 +7,6 @@ import { koa_jwt_token_decoded_ } from './koa_jwt_token_decoded_.js'
  * @private
  */
 export async function verify_jwt_user_id(ctx, authorization) {
-	const jwt_token_decoded = await koa_jwt_token_decoded_(ctx, authorization)
+	const jwt_token_decoded = await strict_verify_jwt_token(ctx, authorization)
 	return user_id_(jwt_token_decoded)
 }